+++ src/App.jsx
@@ -1,11 +1,16 @@
 import './App.css'
+import Card from "./components/card/Card.jsx";
 
 function App() {
 
 
 
   return (
-    <>
+
+      <>
+      <button onClick={Card}> click me to see some pokis </button>
+
+
       <h1>Gotta catch em all!</h1>
     </>
   )
